# api-gateway/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.CheckDocumentRequest:
    properties:
      error:
        type: string
      status:
        type: string
    type: object
  github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.CreateDocumentRequest:
    properties:
      imageType:
        type: string
      objectKey:
        type: string
    type: object
...
  github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.Document:
    properties:
      createdAt:
        type: string
      id:
        type: string
      imageType:
        type: string
      imageURL:
        type: string
      reason:
        type: string
      status:
        type: string
      updatedAt:
        type: string
      userID:
        type: string
    type: object
  github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.DocumentsResponse:
    properties:
      documents:
        items:
          $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.Document'
        type: array
    type: object
  github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorDetail:
    properties:
      message:
        type: string
    type: object
  github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorDetail'
    type: object
  github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.GetUploadDocumentDataRequest:
    properties:
      imageType:
        type: string
    type: object
...
  github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.IDResponse:
    properties:
      id:
        type: string
    type: object
  github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ImageUploadData:
    properties:
      objectKey:
        type: string
      presignedPutURL:
        type: string
    type: object
  github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ImageUploadResponse:
    properties:
      uploadData:
        $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ImageUploadData'
    type: object
...
host: localhost:4000
info:
  contact:
    email: support@carrental.kz
    name: API Support
  description: API Gateway for the Kazakhstan carsharing platform. Handles auth, users,
    fleet, zones, insurance, and maintenance.
  license:
    name: Proprietary
  termsOfService: http://swagger.io/terms/
  title: Car Rental API Gateway
  version: "1.0"
paths:
...
  /users/{id}/documents/processed:
    get:
      description: Returns all processed documents belonging to the specified user.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.DocumentsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get processed documents for a user
      tags:
      - documents
...
  /users/documents:
    post:
      consumes:
      - application/json
      description: Creates a document record after the image has been uploaded to
        object storage.
      parameters:
      - description: Document payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.CreateDocumentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.IDResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create document record
      tags:
      - documents
  /users/documents/check/{id}:
    post:
      consumes:
      - application/json
      description: Sets the review status of a document (e.g. approved or rejected)
        and optionally provides a rejection reason.
      parameters:
      - description: Document ID
        in: path
        name: id
        required: true
        type: string
      - description: Review payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.CheckDocumentRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Review a document
      tags:
      - documents
  /users/documents/upload:
    post:
      consumes:
      - application/json
      description: Returns a presigned PUT URL for uploading a document image to object
        storage.
      parameters:
      - description: Image type
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.GetUploadDocumentDataRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ImageUploadResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get document upload URL
      tags:
      - documents
...
  /users/profile/image-upload:
    get:
      description: Returns a presigned PUT URL and object key for uploading the current
        user's profile image to object storage.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ImageUploadResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get profile image upload URL
      tags:
      - documents
...
securityDefinitions:
  BearerAuth:
    description: 'Enter: Bearer <access_token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"